Update InfoCardHeader design
